Backend/requirements.txt
Flask==2.2.5
Werkzeug==3.0.2
langgraph==0.4.7
pymilvus==2.5.7
python-dotenv==1.0.1
llama-cloud==0.1.22
llama-cloud-services==0.6.23
cohere==5.15.0
langchain-core==0.3.55
langchain-community==0.3.22
langchain-groq==0.3.2
ipython==8.10.0